libusbg: Extract internal header file
